+++ src/sql/relation.rs
@@ -27,7 +27,7 @@ use std::{
     iter::{once, Iterator},
     rc::Rc,
     result,
-    str::FromStr,
+    str::FromStr, fmt::format,
 };
 
 /*