+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Protection operation
+- Set operations support
 
+### Fixed
 ## [0.1.1] - 2023-05-26
 ### Changed
 - Made sqlite optional
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2023-05-25
 ### Added
 - First commit open source
+
 ### Fixed
 ### Changed
 ### Removed